Drain packets, and always stall on input full
It was recommended to me that I drain all packets until no more are
available. So I put them in an std::deque.

Additionally, I got rid of the timeout because it's really not necessary
and probably being triggered unintentionally under certain
circumstances. It's perfectly fine if the encoder can't keep up anyway.